test(acpx): split timestamp contract coverage
